small api proposal for waves

+++ core/src/me/minidigger/projecttd/components/HealthComponent.java
@@ -8,7 +8,7 @@ import com.badlogic.gdx.utils.Pool;
  */
 public class HealthComponent implements Component, Pool.Poolable {
 
-    public double health = 100;
+    public float health = 100;
 
     @Override
     public void reset() {
